Add more rbls
Made the query tolerate errors better. If the service is down or
something goes wrong, we assume that the ip is not on the rbl.

Note that we can make this query much nicer still. It could be nice to
return the names of the lists on which the ip is rather than returning a
boolean result. At least this is a start.